ajustando migrations

+++ backend/src/main/resources/db/migration/V4__criar_tabela_foto.sql
@@ -3,6 +3,6 @@ CREATE TABLE foto
     id        BIGINT AUTO_INCREMENT PRIMARY KEY,
     nome      VARCHAR(255) NOT NULL,
     url       VARCHAR(255) NOT NULL,
-    preco DOUBLE NOT NULL,
+    preco     DECIMAL(10,2) NOT NULL,
     descricao VARCHAR(255)
 );
